refactor: UserFixture 패키지 경로 변경
