+++ src/test/java/com/jj/swm/domain/studyroom/service/StudyRoomCommandServiceIntegrationTest.java
@@ -8,7 +8,7 @@ import com.jj.swm.domain.studyroom.entity.StudyRoomReserveType;
 import com.jj.swm.domain.studyroom.fixture.StudyRoomFixture;
 import com.jj.swm.domain.studyroom.entity.StudyRoom;
 import com.jj.swm.domain.studyroom.repository.*;
-import com.jj.swm.domain.user.UserFixture;
+import com.jj.swm.domain.user.fixture.UserFixture;
 import com.jj.swm.domain.user.entity.User;
 import com.jj.swm.domain.user.repository.UserRepository;
 import com.jj.swm.global.exception.GlobalException;
